requirement.txt
Django==4.2
Pillow==9.5.0
PyJWT==2.6.0
PyYAML==6.0
asgiref==3.6.0
beautifulsoup4==4.12.2
django-bootstrap4==23.1
django-cors-headers==3.14.0
django-crispy-forms==1.10.0
django-uni-form==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
helper==2.5.0
layout==0.4.1
mysqlclient==2.1.1
pip==23.2.1
pytz==2023.3
setuptools==68.0.0
soupsieve==2.4.1
sqlparse==0.4.4